requirement.txt
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
fonttools==4.43.1
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
kiwisolver==1.4.5
matplotlib==3.8.1
moviepy==1.0.3
numpy==1.26.1
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.0.7
